Added def user input, still working on fleshing out automatic floorplan creation

+++ Makefile
@@ -12,7 +12,7 @@ extra               ?=  # extra configs
 args                ?=  # command-line args (including step flow control)
 
 
-OBJ_DIR             ?= $(vlsi_dir)/build-$(pdk)-$(tools)/$(design)
+OBJ_DIR             ?= $(vlsi_dir)/build-$(pdk)-$(tools)/$(design)-effort_high
 
 # non-overlapping default configs
 ENV_YML             ?= $(vlsi_dir)/configs-env/$(env)-env.yml
@@ -35,7 +35,7 @@ DESIGN_SYN_CONF     ?= configs-design/$(design)/syn.yml
 DESIGN_PAR_CONF     ?= configs-design/$(design)/par.yml
 
 
-PROJ_YMLS           ?= $(PDK_CONF) $(TOOLS_CONF) $(DESIGN_CONF) $(DESIGN_PDK_CONF) $(DESIGN_PAR_CONF) $(SIM_CONF) $(POWER_CONF) $(SRAM_CONF) $(extra)
+PROJ_YMLS           ?= $(PDK_CONF) $(TOOLS_CONF) $(DESIGN_CONF) $(DESIGN_PDK_CONF) $(DESIGN_SYN_CONF) $(DESIGN_PAR_CONF) $(SIM_CONF) $(POWER_CONF) $(SRAM_CONF) $(extra)
 HAMMER_EXTRA_ARGS   ?= $(foreach conf, $(PROJ_YMLS), -p $(conf)) $(args)
 
 HAMMER_D_MK = $(OBJ_DIR)/hammer.d
@@ -47,10 +47,14 @@ $(HAMMER_D_MK):
 
 -include $(HAMMER_D_MK)
 
+stp:
+	hammer-vlsi -e /bwrcq/scratch/caderichard/hammer_work/configs-env/bwrc-env.yml -p /bwrcq/scratch/caderichard/hammer_work/configs-design/asic-lab/syn.yml -p /bwrcq/scratch/caderichard/hammer_work/build-asap7-cm/asic-lab/syn-rundir/syn-output.json $(HAMMER_EXTRA_ARGS) -o /bwrcq/scratch/caderichard/hammer_work/build-asap7-cm/asic-lab/par-input.json --obj_dir /bwrcq/scratch/caderichard/hammer_work/build-asap7-cm/asic-lab syn-to-par
+
+
 srams:
 	hammer-vlsi -e $(ENV_YML) -p $(TOOLS_CONF) -p $(DESIGN_SYN_CONF) -p $(PDK_CONF) --obj_dir $(OBJ_DIR) sram_generator
 	cp output.json $(SRAM_CONF)
 
 clean:
-	rm -rf $(OBJ_DIR)
+	[ -n "$(OBJ_DIR)" ] && rm -rf $(OBJ_DIR)
 	find $(vlsi_dir) -name 'hammer-vlsi-*.log' -delete
